refactor(homepage): card view for the browser, rename it Detailed View
Follow-up on the lobby browser: it now renders the same map cards as the
homepage instead of a bespoke table. The card markup moves to
components/LobbyCard.ts and both the homepage and the browser call it, so
there is one card to maintain and the two views can't drift. The map
aspect-ratio cache moves with it and is now shared, so a map's manifest is
fetched once per session rather than once per component.

The homepage row is rebalanced: Solo spans two columns and the browser
button is one, matching Join Lobby beneath it. The button and page are
called "Detailed View", and the component, page id, route and i18n keys
were renamed to match.

The browser's own dropdowns now use the settings selects' styling
(appearance-none over an explicit dark background, with our own chevron),
extracted to components/ui/StyledSelect.ts, so they render correctly in
the dark theme instead of falling back to the native control.

The filter panel is regrouped into labelled sections — lobby type and
source, then players-per-team and number-of-teams, then the three ranges
— and the Filters button shows how many filters are active.
