+++ index.html
@@ -305,11 +305,11 @@
           inline
           class="hidden w-full h-full page-content relative z-50"
         ></ranked-modal>
-        <more-games-modal
-          id="page-more-games"
+        <detailed-view-modal
+          id="page-detailed-view"
           inline
           class="hidden w-full h-full page-content relative z-50"
-        ></more-games-modal>
+        ></detailed-view-modal>
       </main-layout>
 
       <div class="[.in-game_&]:hidden mt-auto flex flex-col shrink-0">
