+++ resources/lang/en.json
@@ -462,6 +462,44 @@
     "verified_name": "Verified username",
     "verified_name_info": "Your username is reserved for your account and shows a verified badge in-game."
   },
+  "detailed_view": {
+    "capacity": "Capacity",
+    "delete_profile": "Delete",
+    "filters": "Filters",
+    "hide_empty": "Hide empty lobbies",
+    "joined_players": "Joined players",
+    "layout_duos": "Duos",
+    "layout_quads": "Quads",
+    "layout_trios": "Trios",
+    "max": "Max",
+    "min": "Min",
+    "mode": "Lobby type",
+    "mode_ffa": "FFA",
+    "mode_hvn": "Humans vs Nations",
+    "mode_teams": "Teams",
+    "no_lobbies": "No lobbies match your filters.",
+    "no_profile": "No profile",
+    "per_team": "Players per team",
+    "profile_name": "Profile name",
+    "profiles": "Profiles",
+    "queued": "Up next",
+    "reset": "Reset",
+    "save_profile": "Save",
+    "showing": "Showing {shown} of {total} lobbies",
+    "sort": "Sort",
+    "sort_capacity_asc": "Capacity: low to high",
+    "sort_capacity_desc": "Capacity: high to low",
+    "sort_map_asc": "Map: A to Z",
+    "sort_players_asc": "Players: low to high",
+    "sort_players_desc": "Players: high to low",
+    "sort_starts_soonest": "Starting soonest",
+    "source": "Source",
+    "source_hosted": "Hosted",
+    "source_public": "Public",
+    "team_count": "Number of teams",
+    "team_size": "Team size",
+    "title": "Detailed View"
+  },
   "difficulty": {
     "difficulty": "Nation difficulty",
     "easy": "Easy",
@@ -1006,6 +1044,7 @@
     "clans": "Clans",
     "copyright": "© OpenFront™ and Contributors",
     "create": "Create Lobby",
+    "detailed_view": "Detailed View",
     "discord_avatar_alt": "Discord profile avatar",
     "github": "GitHub",
     "go_to_troubleshooting": "Go to our troubleshooting page",
@@ -1016,7 +1055,6 @@
     "login_discord": "Login with Discord",
     "login_google": "Login with Google",
     "menu": "Menu",
-    "more_games": "More Games",
     "news": "News",
     "play": "Play",
     "privacy_policy": "Privacy Policy",
@@ -1221,45 +1259,6 @@
     "teams_count": "{teamCount} teams",
     "teams_of": "{teamCount} teams of {playersPerTeam}",
     "teams_title": "Teams"
-  },
-  "more_games": {
-    "capacity": "Capacity",
-    "column_map": "Map",
-    "column_mode": "Mode",
-    "column_players": "Players",
-    "column_starts": "Starts",
-    "delete_profile": "Delete",
-    "filters": "Filters",
-    "hide_empty": "Hide empty lobbies",
-    "join": "Join",
-    "joined_players": "Joined",
-    "max": "Max",
-    "min": "Min",
-    "mode": "Lobby type",
-    "mode_ffa": "FFA",
-    "mode_hvn": "Humans vs Nations",
-    "mode_teams": "Teams",
-    "no_lobbies": "No lobbies match your filters.",
-    "no_profile": "No profile",
-    "profile_name": "Profile name",
-    "profiles": "Profiles",
-    "queued": "Up next",
-    "reset": "Reset",
-    "save_profile": "Save",
-    "showing": "Showing {shown} of {total} lobbies",
-    "sort": "Sort",
-    "sort_capacity_asc": "Capacity: low to high",
-    "sort_capacity_desc": "Capacity: high to low",
-    "sort_map_asc": "Map: A to Z",
-    "sort_players_asc": "Players: low to high",
-    "sort_players_desc": "Players: high to low",
-    "sort_starts_soonest": "Starting soonest",
-    "source": "Source",
-    "source_hosted": "Hosted",
-    "source_public": "Public",
-    "team_config": "Team layout",
-    "team_size": "Team size",
-    "title": "More Games"
   },
   "multi_tab": {
     "detected": "Multiple open game tabs detected.",
