+++ src/client/GameModeSelector.ts
@@ -11,17 +11,15 @@ import {
 } from "../core/game/Game";
 import { PublicGameInfo, PublicGames } from "../core/Schemas";
 import "./components/IOSAddToHomeScreenBanner";
+import { lobbyCard, mapAspectRatios } from "./components/LobbyCard";
 import { HostLobbyModal } from "./HostLobbyModal";
 import { JoinLobbyModal } from "./JoinLobbyModal";
 import { PublicLobbySocket } from "./LobbySocket";
 import { JoinLobbyEvent } from "./Main";
 import { SinglePlayerModal } from "./SinglePlayerModal";
-import { terrainMapFileLoader } from "./TerrainMapFileLoader";
 import { UsernameInput } from "./UsernameInput";
 import {
   calculateServerTimeOffset,
-  getMapName,
-  getModifierLabels,
   getSecondsUntilServerTimestamp,
   renderDuration,
   translateText,
@@ -32,7 +30,6 @@ const CARD_BG = "bg-surface";
 @customElement("game-mode-selector")
 export class GameModeSelector extends LitElement {
   @state() private lobbies: PublicGames | null = null;
-  @state() private mapAspectRatios: Map<GameMapType, number> = new Map();
   @state() private inputValid: boolean = true;
   private serverTimeOffset: number = 0;
   private defaultLobbyTime: number = 0;
@@ -99,26 +96,9 @@ export class GameModeSelector extends LitElement {
 
     const allGames = Object.values(lobbies.games ?? {}).flat();
     for (const game of allGames) {
-      const mapType = game.gameConfig?.gameMap as GameMapType;
-      if (mapType && !this.mapAspectRatios.has(mapType)) {
-        // New Map reference triggers Lit reactivity; placeholder ratio 1 lets
-        // has() guard against duplicate in-flight fetches.
-        this.mapAspectRatios = new Map(this.mapAspectRatios).set(mapType, 1);
-        terrainMapFileLoader
-          .getMapData(mapType)
-          .manifest()
-          .then((m: any) => {
-            if (m?.map?.width && m?.map?.height) {
-              this.mapAspectRatios = new Map(this.mapAspectRatios).set(
-                mapType,
-                m.map.width / m.map.height,
-              );
-            }
-          })
-          .catch((e) =>
-            console.error(`Failed to load manifest for ${mapType}`, e),
-          );
-      }
+      mapAspectRatios.ensure(game.gameConfig?.gameMap as GameMapType, () =>
+        this.requestUpdate(),
+      );
     }
   }
 
@@ -129,21 +109,21 @@ export class GameModeSelector extends LitElement {
 
     return html`
       <div class="flex flex-col gap-4 w-full px-4 sm:px-0 mx-auto pb-4 sm:pb-0">
-        <!-- Solo + more games: mobile only, top. Solo keeps the same width as
-             the create/ranked/join buttons below it. -->
+        <!-- Solo + detailed view: mobile only, top. The lobby browser is one
+             column wide, matching Join Lobby below it. -->
         <div class="sm:hidden grid grid-cols-3 gap-4 h-14">
-          ${this.renderSmallActionCard(
-            translateText("main.solo"),
-            this.openSinglePlayerModal,
-            "bg-malibu-blue hover:bg-aquarius active:bg-malibu-blue/80 hover:scale-y-105 hover:scale-x-[1.01]",
-          )}
           <div class="col-span-2">
             ${this.renderSmallActionCard(
-              translateText("main.more_games"),
-              this.openMoreGames,
-              "bg-surface hover:brightness-[1.08] active:brightness-[0.95] hover:scale-105 hover:shadow-[var(--shadow-action-card-hover)]",
+              translateText("main.solo"),
+              this.openSinglePlayerModal,
+              "bg-malibu-blue hover:bg-aquarius active:bg-malibu-blue/80 hover:scale-y-105 hover:scale-x-[1.01]",
             )}
           </div>
+          ${this.renderSmallActionCard(
+            translateText("main.detailed_view"),
+            this.openDetailedView,
+            "bg-surface hover:brightness-[1.08] active:brightness-[0.95] hover:scale-105 hover:shadow-[var(--shadow-action-card-hover)]",
+          )}
         </div>
         <!-- Create/ranked/join: mobile only, below solo -->
         <div class="sm:hidden grid grid-cols-3 gap-4 h-14">
@@ -218,21 +198,21 @@ export class GameModeSelector extends LitElement {
               </div>
             </div>`}
 
-        <!-- Solo + more games, desktop only. Solo is one column wide (matching
-             create/ranked/join); the lobby browser takes the other two. -->
+        <!-- Solo + detailed view, desktop only. Solo spans two columns; the
+             lobby browser is one, the same width as Join Lobby below it. -->
         <div class="hidden sm:grid grid-cols-3 gap-4 h-14">
-          ${this.renderSmallActionCard(
-            translateText("main.solo"),
-            this.openSinglePlayerModal,
-            "bg-malibu-blue hover:bg-aquarius active:bg-malibu-blue/80 hover:scale-y-105 hover:scale-x-[1.01]",
-          )}
           <div class="col-span-2">
             ${this.renderSmallActionCard(
-              translateText("main.more_games"),
-              this.openMoreGames,
-              "bg-surface hover:brightness-[1.08] active:brightness-[0.95] hover:scale-105 hover:shadow-[var(--shadow-action-card-hover)]",
+              translateText("main.solo"),
+              this.openSinglePlayerModal,
+              "bg-malibu-blue hover:bg-aquarius active:bg-malibu-blue/80 hover:scale-y-105 hover:scale-x-[1.01]",
             )}
           </div>
+          ${this.renderSmallActionCard(
+            translateText("main.detailed_view"),
+            this.openDetailedView,
+            "bg-surface hover:brightness-[1.08] active:brightness-[0.95] hover:scale-105 hover:shadow-[var(--shadow-action-card-hover)]",
+          )}
         </div>
         <!-- Bottom row: create + ranked + join (desktop only) -->
         <div class="hidden sm:grid grid-cols-3 gap-4 h-14">
@@ -266,9 +246,9 @@ export class GameModeSelector extends LitElement {
     window.showPage?.("page-ranked");
   };
 
-  private openMoreGames = () => {
+  private openDetailedView = () => {
     if (!this.validateUsername()) return;
-    window.showPage?.("page-more-games");
+    window.showPage?.("page-detailed-view");
   };
 
   private openSinglePlayerModal = () => {
@@ -324,13 +304,6 @@ export class GameModeSelector extends LitElement {
     lobby: PublicGameInfo,
     titleContent: string | TemplateResult,
   ) {
-    const mapType = lobby.gameConfig!.gameMap as GameMapType;
-    const mapImageSrc = terrainMapFileLoader.getMapData(mapType).webpPath;
-    const aspectRatio = this.mapAspectRatios.get(mapType);
-    // Use object-contain for extreme aspect ratios (e.g. Amazon River ~20:1) so
-    // the full map is visible instead of being cropped by object-cover.
-    const useContain =
-      aspectRatio !== undefined && (aspectRatio > 4 || aspectRatio < 0.25);
     const timeRemaining = lobby.startsAt
       ? getSecondsUntilServerTimestamp(lobby.startsAt, this.serverTimeOffset)
       : undefined;
@@ -346,98 +319,14 @@ export class GameModeSelector extends LitElement {
       timeDisplayUppercase = true;
     }
 
-    const mapName = getMapName(lobby.gameConfig?.gameMap);
-
-    const modifierLabels = getModifierLabels(
-      lobby.gameConfig?.publicGameModifiers,
-      lobby.gameConfig?.doomsdayClock?.speed,
-    );
-    // Sort by length for visual consistency (shorter labels first)
-    if (modifierLabels.length > 1) {
-      modifierLabels.sort((a, b) => a.length - b.length);
-    }
-
-    return html`
-      <button
-        @click=${() => this.validateAndJoin(lobby)}
-        ?disabled=${!this.inputValid}
-        class="group relative w-full h-44 sm:h-full text-white uppercase rounded-2xl transition-all duration-200 hover:scale-[1.02] active:scale-[0.98] bg-surface hover:shadow-[var(--shadow-lobby-card-hover)] ${!this
-          .inputValid
-          ? "opacity-50 cursor-not-allowed pointer-events-none"
-          : ""}"
-      >
-        <!-- Image clipped separately so overflow-hidden doesn't block absolute children -->
-        <div
-          class="absolute inset-0 rounded-2xl overflow-hidden pointer-events-none"
-        >
-          ${mapImageSrc
-            ? html`<img
-                src="${mapImageSrc}"
-                alt="${mapName ?? lobby.gameConfig?.gameMap ?? "map"}"
-                draggable="false"
-                class="absolute inset-0 w-full h-full ${useContain
-                  ? "object-contain"
-                  : "object-cover object-center scale-[1.05]"} [image-rendering:auto]"
-              />`
-            : null}
-        </div>
-        <!-- Top row: modifiers + timer -->
-        <div
-          class="absolute inset-x-2 top-2 flex items-start justify-between gap-2"
-        >
-          ${modifierLabels.length > 0
-            ? html`<div class="flex flex-col items-start gap-1 mt-[2px]">
-                ${modifierLabels.map(
-                  (label) =>
-                    html`<span
-                      class="px-2 py-1 rounded text-xs font-bold uppercase tracking-widest bg-malibu-blue text-white shadow-[var(--shadow-malibu-blue-pill)]"
-                      >${label}</span
-                    >`,
-                )}
-              </div>`
-            : html`<div></div>`}
-          <div class="shrink-0">
-            <span
-              class="text-xs font-bold tracking-widest ${timeDisplayUppercase
-                ? "uppercase"
-                : "normal-case"} bg-malibu-blue text-white px-2 py-1 rounded"
-              >${timeDisplay}</span
-            >
-          </div>
-        </div>
-        <!-- Bottom bar: map name + mode, with player count floating above -->
-        <div
-          class="absolute bottom-0 left-0 right-0 flex flex-col px-3 py-2 bg-black/55 backdrop-blur-sm rounded-b-2xl"
-          style="overflow: visible;"
-        >
-          <span
-            class="absolute bottom-full right-2 mb-1 flex items-center gap-1 text-xs font-bold tracking-widest bg-black/70 backdrop-blur-sm px-2 py-0.5 rounded"
-          >
-            ${lobby.numClients}/${lobby.gameConfig?.maxPlayers}
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              class="h-4 w-4 inline-block"
-              viewBox="0 0 20 20"
-              fill="currentColor"
-            >
-              <path
-                d="M13 6a3 3 0 11-6 0 3 3 0 016 0zM18 8a2 2 0 11-4 0 2 2 0 014 0zM14 15a4 4 0 00-8 0v3h8v-3zM6 8a2 2 0 11-4 0 2 2 0 014 0zM16 18v-3a5.972 5.972 0 00-.75-2.906A3.005 3.005 0 0119 15v3h-3zM4.75 12.094A5.973 5.973 0 004 15v3H1v-3a3 3 0 013.75-2.906z"
-              ></path>
-            </svg>
-          </span>
-          ${mapName
-            ? html`<p
-                class="text-sm sm:text-base font-bold uppercase tracking-wider text-left leading-tight"
-              >
-                ${mapName}
-              </p>`
-            : ""}
-          <h3 class="text-xs text-white/70 uppercase tracking-wider text-left">
-            ${titleContent}
-          </h3>
-        </div>
-      </button>
-    `;
+    return lobbyCard({
+      lobby,
+      subtitle: titleContent,
+      timeDisplay,
+      timeDisplayUppercase,
+      disabled: !this.inputValid,
+      onClick: () => this.validateAndJoin(lobby),
+    });
   }
 
   private validateAndJoin(lobby: PublicGameInfo) {
